backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
typing
pymongo
certifi
Pillow
torch
transformers
python-multipart
google-genai
sentence-transformers
cohere
rembg
onnxruntime==1.15.1